#1698: split routing.go into per-domain managers behind a façade
pkg/routing/routing.go was 2085 LOC bundling unrelated control-plane
domains (VRF, route reads + Junos formatters, GRE/IPIP tunnels +
keepalive, XFRM interfaces, next-table/rib-group/PBR ip-rule
reconcilers, bonds, RETH cleanup, interface-monitor status) behind one
*Manager with a single shared ifaceMu. This re-attempt of the
#1544-killed split implements the successor design the #1544 kill
comment prescribed: real per-domain managers with narrow netlink-facing
interfaces, not lines-N-to-M file motion.

Structure:
  - Manager (routing.go, 215 LOC) is now a thin façade: it owns the sole
    *netlink.Handle and delegates every public method (26 methods + 5
    exported free fns, all byte-identical signatures) to the owning
    domain. Zero call-site edits outside pkg/routing/.
  - vrfManager (vrf.go), routeReader (routes.go), tunnelManager
    (tunnel.go), xfrmManager (xfrm.go), nextTable/ribGroup/pbr managers
    (rules.go), bondManager (bond.go), rethManager (reth.go),
    monitorManager (monitor.go) each own their sub-state + their own
    mutex and borrow a narrow ops interface (vrfOps, routeLister,
    linkOps, ruleOps) rather than a concrete handle.

Behavior preserved:
  - The shared ifaceMu is split into per-domain locks. Verified safe: no
    method held ifaceMu across more than one of the disjoint interface
    domains {tunnels, xfrmis, bonds}; keepalives belong to the tunnel
    domain. The dropped cross-domain exclusion is not load-bearing
    (daemon apply is sequential; the netlink socket serializes per
    request).
  - ApplyTunnels' cross-domain call to BindInterfaceToVRF is modeled as
    a tunnelManager.vrfBinder dependency; BindInterfaceToVRF takes no
    lock so there is no lock-ordering cycle.
  - Close() drains keepalive goroutines via tunnel.stopAll() before
    closing the handle (#848). VRF orphan-reap / ownership semantics
    (#844/#847) and rule-reconcile clear-then-add semantics are moved
    verbatim.

Tests: fakeVRFOps gains LinkSetMaster (now part of vrfOps);
TestInterfaceMonitorStatuses constructs monitorManager directly instead
of reaching into Manager internals. All existing pkg/routing tests pass;
whole repo builds clean.
